+++ config/routes.rb
@@ -7,4 +7,5 @@ Pay::Engine.routes.draw do
   post "webhooks/paddle_billing", to: "pay/webhooks/paddle_billing#create" if Pay::PaddleBilling.enabled?
   post "webhooks/paddle_classic", to: "pay/webhooks/paddle_classic#create" if Pay::PaddleClassic.enabled?
   post "webhooks/lemon_squeezy", to: "pay/webhooks/lemon_squeezy#create" if Pay::LemonSqueezy.enabled?
+  post "webhooks/square", to: "pay/webhooks/square#create" if Pay::Square.enabled?
 end