+++ lib/pay/engine.rb
@@ -39,6 +39,7 @@ module Pay
       Pay::PaddleBilling.configure_webhooks if Pay::PaddleBilling.enabled?
       Pay::PaddleClassic.configure_webhooks if Pay::PaddleClassic.enabled?
       Pay::LemonSqueezy.configure_webhooks if Pay::LemonSqueezy.enabled?
+      Pay::Square.configure_webhooks if Pay::Square.enabled?
     end
 
     config.to_prepare do
@@ -46,6 +47,7 @@ module Pay
       Pay::Braintree.setup if Pay::Braintree.enabled?
       Pay::PaddleBilling.setup if Pay::PaddleBilling.enabled?
       Pay::LemonSqueezy.setup if Pay::LemonSqueezy.enabled?
+      Pay::Square.setup if Pay::Square.enabled?
     end
 
     # Determines if a gem version matches requirements